+++ lang/lang_test.go
@@ -1,7 +1,6 @@
 package lang
 
 import (
-	"os"
 	"path"
 	"runtime"
 	"testing"
@@ -12,7 +11,6 @@ import (
 
 type LangTestSuite struct {
 	suite.Suite
-	previousEnv string
 }
 
 func loadTestLang(lang string) {
@@ -21,17 +19,12 @@ func loadTestLang(lang string) {
 }
 
 func (suite *LangTestSuite) SetupSuite() {
-	suite.previousEnv = os.Getenv("GOYAVE_ENV")
-	os.Setenv("GOYAVE_ENV", "test")
 	LoadDefault()
 	LoadAllAvailableLanguages()
 
-	// FIXME workaround waiting for load from path
-	os.Chdir("..")
-	if err := config.Load(); err != nil {
+	if err := config.LoadFrom("../config.test.json"); err != nil {
 		suite.FailNow(err.Error())
 	}
-	os.Chdir("lang")
 	config.Set("app.defaultLanguage", "en-US")
 }
 
@@ -147,7 +140,6 @@ func (suite *LangTestSuite) TestPlaceholders() {
 
 func (suite *LangTestSuite) TearDownAllSuite() {
 	languages = map[string]language{}
-	os.Setenv("GOYAVE_ENV", suite.previousEnv)
 }
 
 func TestLangTestSuite(t *testing.T) {